requirements.txt
pydantic~=1.9.0
setuptools~=57.0.0
kafka-python~=2.0.2
git+https://github.com/doctor3030/loguru-logger.git